add whatsapp status

+++ database/migrations/2023_12_05_183504_add_use_logo_to_settings_table.php
@@ -13,6 +13,10 @@ class AddUseLogoToSettingsTable extends Migration
      */
     public function up()
     {
+        if (Schema::hasColumn('settings', 'key')) {
+            return;
+        }
+
         Schema::table('settings', function (Blueprint $table) {
             $table->integer('use_logo')->default(1);
         });
@@ -25,6 +29,10 @@ class AddUseLogoToSettingsTable extends Migration
      */
     public function down()
     {
+        if (Schema::hasColumn('settings', 'key')) {
+            return;
+        }
+
         Schema::table('settings', function (Blueprint $table) {
             $table->dropColumn('use_logo');
         });
